plugin/requirements.txt
io_scene_gltf2
numpy
fake-bpy-module-4.1